relance_patient/requirements.txt
aiomysql==0.0.20
appnope==0.1.0
asgiref==3.2.10
astroid==2.4.2
async-timeout==3.0.1
attrs==20.2.0
backcall==0.2.0
cairocffi==1.1.0
CairoSVG==2.4.2
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
cryptography==3.2
cssselect2==0.3.0
decorator==4.4.2
defusedxml==0.6.0
Django==3.0.9
django-glrm==1.1.3
djangorestframework==3.11.1
et-xmlfile==1.0.1
gunicorn==20.0.4
h11==0.8.1
h2==3.2.0
hpack==3.0.0
html5lib==1.1
http3==0.6.7
hyperframe==5.2.0
idna==2.10
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
isort==4.3.21
jdcal==1.4.1
jedi==0.17.2
jupyter-client==6.1.7
jupyter-core==4.6.3
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==4.7.6
mysqlclient==2.0.1
numpy==1.19.2
openpyxl==3.0.5
pandas==1.1.
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
prompt-toolkit==3.0.7
ptyprocess==0.6.0
pycares==3.1.1
pycparser==2.20
Pygments==2.6.1
pylint==2.5.3
PyMySQL==0.9.2
PyPDF2==1.26.0
Pyphen==0.9.5
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.1
pyzmq==19.0.2
reportlab==3.5.49
requests==2.24.0
rfc3986==1.4.0
six==1.15.0
sqlparse==0.3.1
tinycss2==1.0.2
toml==0.10.1
tornado==6.0.4
traitlets==5.0.4
urllib3==1.25.10
wcwidth==0.2.5
WeasyPrint==51
webencodings==0.5.1
wrapt==1.12.1
xhtml2pdf==0.2.4
xlrd==1.2.0
yarl==1.6.0
